PheWAS catalog information

A table listing PheWAS Codes (or phecodes) and their corresponding description, category, exclusion range, and other information. Last manually curated on June 1, 2023.

Usage

pheinfo

Format

## `pheinfo` A data frame with 1,866 rows and 12 columns:

phecode

PheWAS Code (aka phecode)

description

Phecode description

group

Phecode disease category or group

order

A variable ordering by `groupnum` and then numeric `phecode` value, useful for plotting

color

Phecode group color inspired by Okabe-Ito colorblind-friendly palette

sex

Defines `male` or `female` specific phecodes, otherwise `both`

rollup

Indicator for whether `phecode` rolls up

leaf

Indicator for whether `phecode` is a leaf (a subcode of a parent code)

groupnum

A numeric value corresponding to `phecode` `group`

color_original

Original `phecode` `group` color scheme

phecode_exclude_range

Range of phecodes that disqualify as controls for a given `phecode`

phecode_exclude_phenotypes

Description of `phecode_exclude_range` codes

...

Source

Curated version of PheWAS::pheinfo